Comments on RecoMuon/MuonIdentification
