+++ RecoMuon/MuonIdentification/BuildFile.xml
@@ -32,10 +32,12 @@
 <use   name="Geometry/CSCGeometry"/>
 <use   name="Geometry/GEMGeometry"/>
 <use   name="Geometry/Records"/>
+<!--
 <use   name="SimDataFormats/Track"/>
 <use   name="SimDataFormats/TrackingHit"/>
 <use   name="SimDataFormats/Associations"/>
 <use   name="SimDataFormats/TrackingAnalysis"/>
+-->
 <use   name="SimTracker/Records"/>
 <use   name="SimMuon/MCTruth"/>
 <use   name="SimTracker/TrackAssociation"/>
