+++ RecoMuon/MuonIdentification/interface/ME0MuonSelector.h
@@ -6,7 +6,6 @@
 
 #include "FWCore/Framework/interface/ESHandle.h"
 
-#include "TMath.h"
 #include <string>
