+++ RecoMuon/MuonIdentification/plugins/ME0MuonConverter.cc
@@ -33,6 +33,59 @@
 #include "DataFormats/RecoCandidate/interface/RecoChargedCandidate.h"
 #include "DataFormats/RecoCandidate/interface/RecoChargedCandidateFwd.h"
 
+#ifndef EmulatedME0Segment_ME0MuonConverter_h
+#define EmulatedME0Segment_ME0MuonConverter_h
+
+/** \class ME0MuonConverter 
+ * Produces a collection of ME0Segment's in endcap muon ME0s. 
+ *
+ *
+ * \author David Nash
+ */
+
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/stream/EDProducer.h"
+#include "FWCore/Framework/interface/Event.h"
+#include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+
+
+#include "DataFormats/RecoCandidate/interface/RecoChargedCandidate.h"
+#include "DataFormats/RecoCandidate/interface/RecoChargedCandidateFwd.h"
+
+
+#include "FWCore/Framework/interface/EDProducer.h"
+#include "FWCore/Utilities/interface/InputTag.h"
+
+#include <DataFormats/GEMRecHit/interface/ME0SegmentCollection.h>
+
+
+
+#include <DataFormats/MuonReco/interface/ME0Muon.h>
+#include <DataFormats/MuonReco/interface/ME0MuonCollection.h>
+
+
+
+class ME0MuonConverter : public edm::stream::EDProducer<> {
+public:
+  /// Constructor
+  explicit ME0MuonConverter(const edm::ParameterSet&);
+  /// Destructor
+  ~ME0MuonConverter();
+  /// Produce the converted collection
+  virtual void produce(edm::Event&, const edm::EventSetup&) override;
+
+    
+
+private:
+
+  //edm::EDGetTokenT<std::vector<reco::ME0Muon>> OurMuonsToken_;
+  edm::EDGetTokenT<ME0MuonCollection> OurMuonsToken_;
+};
+
+#endif
+
 
 ME0MuonConverter::ME0MuonConverter(const edm::ParameterSet& pas) {
 	
@@ -67,7 +120,7 @@ void ME0MuonConverter::produce(edm::Event& ev, const edm::EventSetup& setup) {
     Particle::LorentzVector p4(tkRef->px(), tkRef->py(), tkRef->pz(), tkRef->p());
     Particle::Point vtx(tkRef->vx(),tkRef->vy(), tkRef->vz());
 
-    int pid = 13;
+    int pid = 0;
     if(abs(q)==1) pid = q < 0 ? 13 : -13;
     reco::RecoChargedCandidate cand(q, p4, vtx, pid);
     cand.setTrack(thisMuon->innerTrack());
